+++ tapps_agents/workflow/agent_handlers/registry.py
@@ -82,6 +82,7 @@ class AgentHandlerRegistry:
         from .debugger_handler import DebuggerHandler
         from .designer_handler import DesignerHandler
         from .documenter_handler import DocumenterHandler
+        from .enhancer_handler import EnhancerHandler
         from .implementer_handler import ImplementerHandler
         from .ops_handler import OpsHandler
         from .orchestrator_handler import OrchestratorHandler
@@ -98,6 +99,7 @@ class AgentHandlerRegistry:
             PlannerHandler,
             ArchitectHandler,
             DesignerHandler,
+            EnhancerHandler,
             OpsHandler,
             DocumenterHandler,
             OrchestratorHandler,